Fix bug where git tries to use feature branch for tag.
You can't have `/` in a git tag - what I need to do is replace the
`$GIT_BRANCH` env variable with another one called `$DEPLOY_ENV` which
can be either `dev` or `main` but is specified only from the outside,
not from the git branch. Using the git branch was a bad idea. I don't
have time to update this now, though.